Corrigiendo el tipo de dato en la configuración de alertas en el componente alert-config. Se cambia el método de parseo de 'parseInt' a 'parseFloat' para asegurar que los valores de alerta se manejen correctamente como números decimales, mejorando la precisión en la configuración de alertas.

+++ src/app/modules/generador/components/inputs-generator/alert-config/alert-config.component.ts
@@ -179,7 +179,7 @@ export class AlertConfigComponent implements OnInit, OnChanges {
       if (!key.startsWith('planesCuidado_') && formValue[key]) {
         // Configuración base con el valor de la alerta
         configuracion[key] = {
-          valor: parseInt(formValue[key])
+          valor: parseFloat(formValue[key])
         };
 
         // Solo agregar planes de cuidado si existen y no están vacíos
